requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
importlib_metadata==8.2.0
Markdown==3.6
mysqlclient==2.2.4
sqlparse==0.5.1
typing_extensions==4.12.2
zipp==3.19.2
